+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/worksheet/excel_comments/excel_comment.rb
@@ -13,7 +13,7 @@ module OoxmlParser
     # @return [ExcelComment] result of parsing
     def parse(node)
       node.xpath('xmlns:text/xmlns:r').each do |node_child|
-        @characters << ParagraphRun.new(parse: self).parse(node_child)
+        @characters << ParagraphRun.new(parent: self).parse(node_child)
       end
       self
     end