+++ Changelog.md
@@ -58,6 +58,7 @@
 * Do not crash, just show stderr if resource not found
 * Add base support of `chartsheets`
 * Add `parse_hex_string` for 3 digit colors
+* Add parsing 'DocxParagraph#inserted' context
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit